+++ src/workers/calc-ref-orbit.ts
@@ -16,9 +16,9 @@ import {
   mul,
   mulN,
   norm,
+  pixelToComplexCoordinateComplexArbitrary,
   toComplex,
-} from "../math";
-import { pixelToComplexCoordinateComplexArbitrary } from "../math/complex-plane";
+} from "../math/complex";
 import { BLATableBuffer, RefOrbitWorkerParams, XnBuffer } from "../types";
 
 export type RefOrbitContext = {
